Upgrading the clist file access with 64 bits file access.
DETAILS :

Bug 688394 "clist playback of large pdf image segfaults".

clist file position variable changed type to int64_t.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7041 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gxclist.c
@@ -657,7 +657,7 @@ clist_end_page(gx_device_clist_writer * cldev)
 #ifdef DEBUG
     if (gs_debug_c('l') | gs_debug_c(':'))
 	dlprintf2("[:]clist_end_page at cfile=%ld, bfile=%ld\n",
-		  cb.pos, cldev->page_bfile_end_pos);
+		  (long)cb.pos, (long)cldev->page_bfile_end_pos);
 #endif
     return 0;
 }
